refactor: rename package -> windows
