+++ client.go
@@ -2,7 +2,7 @@ package gowindows
 
 import (
 	"github.com/d-strobel/gowindows/connection"
-	"github.com/d-strobel/gowindows/package/local"
+	"github.com/d-strobel/gowindows/windows/local"
 )
 
 type Client struct {
