+++ windows/local/local_client_test.go
